tree.getEpsilon Start dev compare

+++ src/main/java/Tree/ISNode.java
@@ -3,7 +3,7 @@ package Tree;
 public interface ISNode {
 
     void incQuantity();
-    void setHeight(double height);
+    void setHeight(int height);
     void addChild(SNode child);
     void uplvl();
